+++ examples/src/lib.rs
@@ -1,9 +1,6 @@
 extern crate alloc;
 
-mod format;
 pub mod simple;
 
 #[cfg(test)]
 mod simple_malicious;
-
-pub use format::Bincode;